Resolve confusion about artifacts.repo.username
This username is used for authentication, not just as a name for the
testing branch. It happens to be irrelevant if you are using an access
token for authentication (at least as far as I can tell), but if you
use a password, it's probably more relevant.

Emphasise that it should match your username on the configured code
hosting platform.